Added MessageChannel#editMessageById
